Put particle physics unto cActor loop.
The code for particle physics is now run in the same loop as the cActors. At first, the idea was to minimize the number of loops per frame, but that means the particles are updated four times more frequently, so it is a lot slower at high numbers. However, since the particles behave so much more consistently and beautifully now that they are updated along-side the cActors, i cannot let inefficiency stop me.

Many things must be adjusted.