Replaced type Sticker with Negative.
Though i'm thinking of renaming it to Hunter. It repulses all circle actors (aside from the player), but is heavily attracted to them. Amazingly, an orbital will be repulsed by it, it will be pulled towards that orbital, and innevitably catch up.

It seems that making its pray run inproves its success.