CircleActor::move now move_impl.
The sig for Actor::move is void(int), but the for CircleActor::move it's void(int,float).

llvm did not consider it an overload of Actor::move and called it directly from main instead of Player::move or Orbital::move. Changing it to move_impl doesn't break any code and all circle actors call it instead of Actor::move, so only Orbital::move and Player::move needed to be changed.